Check for score before team delete

+++ server/routers/crud_router.js
@@ -1,9 +1,11 @@
 'use strict'
 const express = require('express')
-const db = require('../utilities/mongo_connection')
+
 const ObjectId = require('mongodb').ObjectID
 const MsLogger = require('@first-lego-league/ms-logger').Logger()
 const { authroizationMiddlware } = require('@first-lego-league/ms-auth')
+
+const db = require('../utilities/mongo_connection')
 
 const mhubConnection = require('../utilities/mhub_connection')
 
@@ -116,16 +118,22 @@ exports.getRouter = function (options) {
     if (!validationResult) {
       res.sendStatus(400)
     }
-    db.connection().then(connection => {
-      connection.db().collection(options.collectionName).findOneAndDelete(idMongoQuery(options.IdField, req.params.id)).then(dbResponse => {
-        if (dbResponse.ok === 1) {
-          mhubConnection.publishUpdateMsg(options.mhubNamespace)
-          res.sendStatus(200)
-        }
+    validationResult.then(error => {
+      if (error && error.name === 'Error') {
+        res.status(500).send(error.message)
+        return
+      }
+      db.connection().then(connection => {
+        connection.db().collection(options.collectionName).findOneAndDelete(idMongoQuery(options.IdField, req.params.id)).then(dbResponse => {
+          if (dbResponse.ok === 1) {
+            mhubConnection.publishUpdateMsg(options.mhubNamespace)
+            res.sendStatus(200)
+          }
+        })
+      }).catch(err => {
+        MsLogger.error(err)
+        res.sendStatus(500)
       })
-    }).catch(err => {
-      MsLogger.error(err)
-      res.sendStatus(500)
     })
   })
